Sound support & script "object" folder changes
Sounds are now moved over into the assets folder with corresponding data from GameMaker. The objects inside of the scripts folder can now be in any subfolder, so long as they reside in the upper "objects" folder. I also added a timer class so that interpolation is an option w/ a fixed timestep currently locked at 60 TPS.