Shader support and work-in-progress sprite padding
Sprite padding works, it just has some unwanted results when dealing with hand-set sampler UVs in Lua when trying to reference, say, another texture. Wanting to implement a function in the sprite index to get the UVs for a frame so that can be passed as a vec4 argument, but I'd hate to have to save state (sol2 state) in that as well.
Curious if there's a way to access the active Lua state from the SpriteIndex userdata functions like how vanilla C Lua provides it as an argument..