Unmanaged music & sound folder support
Unmanaged music & sounds take prescedence over managed files. This makes it easy to override sounds that existed in the original base but should be used carefully as you don't want to forget the originals are there. The idea behind this push is to ultimately allow packs to override data in mods, etc. without having to utilize the GMS IDE, sans tilemaps. Regardless, I don't think anything should be sandboxed into the GM scope and anything added by hand should be done with a level of freedom and ease so that if one so chooses, the maximum you'd need to add in GMS is just the sprites/objects for the visuals used in the tilemaps!