Easier to discern BaseObject type; sound deferring
Since objects and instances share classes, the codebase could be a bit convoluted as a result and tedious to understand some functions at times. Doing this clarifies the separation of instance objects and base class objects far better for me.

Sounds are now loaded on a separate thread so not to cause a stutter when one is used for the first time. When using the new profiler, I noticed that a frame could take upwards of 10-15 milliseconds when a sound was used for the first time, so this entirely negates that at the cost of the first sound potentially playing a bit later than when it should actuate. However, given that the time period is menial, it's not noticable, and this is an average fix to this problem (see Minecraft).

Also added a simple debugging profiler. This is just a quick and dirty way to help me find the pain points in code that may not be performing how I want them to.