Sound engine re-work (WIP); Lua crash fix
The previous sound engine was getting a bit cumbersome, and with my movement towards ambiguous ownership on the game side where the user does not need to keep track of object lifetimes, I deemed it fitting to rework what was there. It's also far easier to read.

The only thing left to do is to re-evaluate how the initial sound(s) loads in if it takes a moment on older hardware, and how to keep track of said instances in a multithreaded scenario. Deferred sound creation with a queued instance list is likely the answer in this case

I enabled all sol2 safeties to the compiler definitions and was able to weed out some minor casting oversights as well, which although did not cause any visible problems, will prevent future problems from arising as the engine should never be the reason the game crashes.